requirements.txt
streamlit==1.36.0

agno==2.1.10
groq==0.9.0

PyGithub==2.3.0
requests==2.32.3

tree-sitter==0.21.3
tree-sitter-languages==1.10.2

faiss-cpu==1.8.0
numpy==1.26.4

python-dotenv==1.0.1
pydantic==2.8.2
typing-extensions==4.12.2
rich==13.7.1

tiktoken==0.7.0
httpx==0.27.0